Title: 낮은 지점들, Time: 69ms, Memory: 2MB, Status: Passed - CodeTree

+++ 250322/낮은 지점들/lowest-points.cpp
@@ -22,7 +22,7 @@ int main(){
         }
     }
 
-    int answer = 0;
+    long long answer = 0;
     for (auto &dot : m){
         answer += dot.second;
     }
